requirements.txt
torch>=1.8.0
ultralytics-thop>=2.0.0
pillow>=7.1.2
tqdm>=4.64.0
opencv-python>=4.6.0
numpy>=1.23.0,<=2.1.1
matplotlib>=3.3.0
requests>=2.23.0
seaborn>=0.11.0
pandas>=1.1.4
scipy>=1.4.1
psutil
pyyaml>=5.3.1
torchvision>=0.9.0
py-cpuinfo
